services/django-socketio/Scripts/requirements.txt
python-socketio
djangorestframework
eventlet
pyjwt